before 5/26/23

+++ minMax.cs
@@ -15,12 +15,17 @@ namespace TestingConsoleApp
 
 		public static void miniMaxSum(List<int> arr)
 		{
+			//used these for test passing;
+			//long minSum = long.MaxValue, maxSum = long.MinValue;
+			//minSum = Math.Min(minSum, sum);
+			//maxSum = Math.Max(maxSum, sum);
+
 			long max = 0;
 			long min = 0;
 			int len = arr.Count;
 			for (int a = 0; a < len; a++)
 			{
-				long tempTotal = 0;
+				int tempTotal = 0;
 				for (int b = 0; b < len; b++)
 				{
 					if (a != b)
@@ -36,9 +41,13 @@ namespace TestingConsoleApp
 				{
 					min = tempTotal;
 				}
-
 			}
 			Console.WriteLine(min + " " + max);
+			// using linq but in this case failing overflow
+			//var arr2 = new List<int> { 89, 7, 99, 4, 0 };
+			//arr2.Sort();
+			//long mnSum = arr2.Take(4).Sum();
+			//long mxSum = arr2.Skip(1).Sum();
 		}
 	}
